even shorter path
