+++ src/main/kotlin/io/quarkus/code/CodeQuarkusResource.kt
@@ -3,7 +3,6 @@ package io.quarkus.code
 import io.quarkus.code.model.CodeQuarkusExtension
 import io.quarkus.code.model.Config
 import io.quarkus.code.model.QuarkusProject
-import io.quarkus.code.model.ShortUrl
 import io.quarkus.code.services.CodeQuarkusConfigManager
 import io.quarkus.code.services.QuarkusExtensionCatalog
 import io.quarkus.code.services.QuarkusProjectCreator
@@ -11,7 +10,6 @@ import org.eclipse.microprofile.openapi.annotations.Operation
 import org.eclipse.microprofile.openapi.annotations.media.Content
 import org.eclipse.microprofile.openapi.annotations.media.Schema
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse
-import java.net.URI
 import javax.inject.Inject
 import javax.validation.Valid
 import javax.ws.rs.BeanParam
@@ -19,10 +17,9 @@ import javax.ws.rs.GET
 import javax.ws.rs.Path
 import javax.ws.rs.Produces
 import javax.ws.rs.core.MediaType.APPLICATION_JSON
-import javax.ws.rs.core.MediaType.TEXT_PLAIN
 import javax.ws.rs.core.Response
 
-@Path("/")
+@Path("/api")
 class CodeQuarkusResource {
 
     @Inject
@@ -61,36 +58,6 @@ class CodeQuarkusResource {
     fun extensions(): List<CodeQuarkusExtension> {
         return extensionCatalog.extensions
     }
-
-    @GET
-    @Path("/shorten")
-    @Produces(TEXT_PLAIN)
-    @Operation(summary = "Create a short url based on the parameters")
-    fun createShort(@Valid @BeanParam project: QuarkusProject): Response {
-        val url = "https://code.quarkus.io/api/download?g=${project.groupId}&a=${project.artifactId}&v=${project.version}&c=${project.className}&e=${project.extensions}"
-        val response = { shortUrl: ShortUrl ->
-            Response.ok("https://code.quarkus.io/api/shorten/${shortUrl.id}").build()
-        }
-        urlRepository.getByUrl(url)?.let { shortUrl ->
-            return response(shortUrl)
-        }
-        val shortUrl = ShortUrl(url = url)
-        urlRepository.save(shortUrl)
-
-        return response(shortUrl)
-    }
-
-    @GET
-    @Path("/shorten/{id}")
-    @Operation(summary = "Redirect user to download for this id")
-    fun getShort(@PathParam("id") id: String): Response {
-        val shortUrl = urlRepository.getById(id)
-        shortUrl?.url?.let { url ->
-            return Response.seeOther(URI(url)).build()
-        }
-        return Response.serverError().build()
-    }
-
 
     @GET
     @Path("/download")
