+++ src/main/resources/application.properties
@@ -1,4 +1,3 @@
-quarkus.resteasy.path=/api
 quarkus.log.file.enable=false
 quarkus.http.cors=true
 io.quarkus.code.quarkus-version=${version.quarkus}
